fix Renderer

+++ FlockingBoids/Renderer/IRenderer.cs
@@ -6,6 +6,6 @@ namespace Renderer
     {
         void Clear(Color color);
         void DrawLine(Point pt1, Point pt2, double lineWidth, Color color);
-        void FillCircle(Point center, double radius, Color color);
+        void FillCircle(Point center, float radius, Color color);
     }
 }
